Move components to a components folder
