+++ client/components/Reviews/components/ReviewsComponent.jsx
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from 'react';
-import Container from './styles/ReviewsComponent.styled';
+
 import Summary from './Summary';
 import Reviews from './Reviews';
-import { sortReviews } from './helpers/helpers';
+import Container from '../styles/ReviewsComponent.styled';
 
-import exampleGetReviews from './sampleData/exampleGetReviews';
-import exampleReviewsMeta from './sampleData/exampleReviewsMeta';
+import { sortReviews } from '../helpers/helpers';
+import exampleGetReviews from '../sampleData/exampleGetReviews';
+import exampleReviewsMeta from '../sampleData/exampleReviewsMeta';
 
 const ReviewsComponent = () => {
   const [sort, setSort] = useState(0);
