+++ client/components/Reviews/tests/ReviewsList.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import ReviewsList from '../ReviewsList';
+import ReviewsList from '../components/ReviewsList';
 import exampleGetReviews from '../sampleData/exampleGetReviews';
 
 const exampleReviews = exampleGetReviews.results;
