+++ client/components/Reviews/tests/Summary.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import Summary from '../Summary';
+import Summary from '../components/Summary';
 import exampleReviewsMeta from '../sampleData/exampleReviewsMeta';
 
 it('should render a summary component', () => {
